--- a/codegen-test/model/rest-json-extras.smithy
+++ b/codegen-test/model/rest-json-extras.smithy
@@ -53,7 +53,16 @@ apply QueryPrecedence @httpRequestTests([
 @restJson1
 service RestJsonExtras {
     version: "2019-12-16",
-    operations: [StringPayload, PrimitiveIntHeader, EnumQuery, StatusResponse, MapWithEnumKeyOp, PrimitiveIntOp]
+    operations: [
+        StringPayload,
+        PrimitiveIntHeader,
+        EnumQuery,
+        StatusResponse,
+        MapWithEnumKeyOp,
+        PrimitiveIntOp,
+        EscapedStringValues,
+        NullInNonSparse,
+    ]
 }
 
 @http(uri: "/StringPayload", method: "POST")
@@ -187,4 +196,70 @@ structure MapWithEnumKeyInputOutput {
 operation MapWithEnumKeyOp {
     input: MapWithEnumKeyInputOutput,
     output: MapWithEnumKeyInputOutput,
+}
+
+
+@enum([
+    { value: "has\"quotes", name: "HAS_QUOTES" },
+    { value: "normal", name: "NORMAL" },
+])
+string EnumWithEscapedChars
+
+structure EscapedStringValuesInputOutput {
+    enum: EnumWithEscapedChars,
+    @jsonName("also\"has\"quotes")
+    someString: String,
+}
+
+@http(uri: "/escaped-string-values", method: "POST")
+@httpRequestTests([
+    {
+        id: "EscapedStringValuesRequest",
+        uri: "/escaped-string-values",
+        method: "POST",
+        protocol: "aws.protocols#restJson1",
+        body: "{\"enum\":\"has\\\"quotes\",\"also\\\"has\\\"quotes\":\"test\"}",
+        params: { enum: "has\"quotes", someString: "test" },
+    }
+])
+@httpResponseTests([
+    {
+        id: "EscapedStringValuesResponse",
+        protocol: "aws.protocols#restJson1",
+        code: 200,
+        body: "{\"enum\":\"has\\\"quotes\",\"also\\\"has\\\"quotes\":\"test\"}",
+        params: { enum: "has\"quotes", someString: "test" },
+    }
+])
+operation EscapedStringValues {
+    input: EscapedStringValuesInputOutput,
+    output: EscapedStringValuesInputOutput,
+}
+
+list NonSparseList {
+    member: String,
+}
+
+map NonSparseMap {
+    key: String,
+    value: String,
+}
+
+structure NullInNonSparseOutput {
+    list: NonSparseList,
+    map: NonSparseMap,
+}
+
+@http(uri: "/null-in-non-sparse", method: "POST")
+@httpResponseTests([
+    {
+        id: "NullInNonSparse",
+        protocol: "aws.protocols#restJson1",
+        code: 200,
+        body: "{\"list\":[null,\"one\",null,\"two\",null],\"map\":{\"zero\":null,\"one\":\"1\"}}",
+        params: { list: ["one", "two"], map: { "one": "1" } },
+    }
+])
+operation NullInNonSparse {
+    output: NullInNonSparseOutput,
 }