--- a/codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parse/XmlBindingTraitParserGenerator.kt
+++ b/codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parse/XmlBindingTraitParserGenerator.kt
@@ -394,12 +394,7 @@ class XmlBindingTraitParserGenerator(
                 *codegenScope, "Shape" to symbol
             ) {
                 Attribute.AllowUnusedMut.render(this)
-                rustTemplate(
-                    """
-                    let mut builder = #{Shape}::builder();
-                """,
-                    *codegenScope, "Shape" to symbol
-                )
+                rustTemplate("let mut builder = #{Shape}::builder();", *codegenScope, "Shape" to symbol)
                 val members = shape.xmlMembers()
                 if (members.isNotEmpty()) {
                     parseStructureInner(members, "builder", Ctx(tag = "decoder", accum = null))