--- a/codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parse/JsonParserGeneratorTest.kt
+++ b/codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parse/JsonParserGeneratorTest.kt
@@ -72,6 +72,7 @@ class JsonParserGeneratorTest {
         }
 
         structure Top {
+            @required
             choice: Choice,
             field: String,
             extra: Integer,