More progress towards #161. This PR implements JSON deserialization codegen using the new smithy-json token streaming parser. Cleaning up the old Serde implementation is going to be a separate PR as it's going to require more refactoring.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.