@jdisanti 2021-06-14
> And seems like `deserializeFunctionName` is generating an extra `_` when called on lists which is causing CI to fail

It looks like this is actually because MediaPackage has names such as `com.amazonaws.mediapackage#__mapOf__string`. I think we just need to suppress this warning.