@jdisanti 2021-06-15
Did some crude benchmarking to get a rough idea of how these changes impact performance. Note: this isn't a completely fair comparison right now because the smithy-json version is still using serde to parse errors (that part hasn't been refactored as part of this PR).

Repeating json-rpc10, json-rpc11, rest-json, and rest-json-extras protocol tests 100x in release mode (wall-clock time):
```
smithy-json: 42.354 seconds
serde-json:  44.656 seconds
```

Dynamo release build time including dependencies (wall-clock time):
```
smithy-json: 1m 36s
serde-json:  1m 29s
```

Running `cargo llvm-lines`:
```
smithy-json: 528,876
serde-json:  760,254
```

I want to see the same comparisons after refactoring the error parsing, and also want to add a real-world Dynamo GetItem comparison.