@rcoh 2021-06-15
big +1 to both of those! I would make the integration test gradle job slightly smarter and allow us to add a `benches` folder. We can eventually even run these benchmarks in CI. I'd also be curious about absolute numbers for serialization