@jdisanti 2021-06-16
Added a DynamoDB benchmark in #507. I went ahead and ran it against both `main` and this PR (after running `./gradlew :aws:sdk:assemble` for each), and the results are not that different:

main:
```
ser_deser_bench         time:   [224.39 us 226.10 us 228.14 us]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```

PR:
```
ser_deser_bench         time:   [226.49 us 227.60 us 228.89 us]
                        change: [-0.2383% +0.5420% +1.3489%] (p = 0.19 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
```

I'm not fully convinced this benchmark is giving me the information I want, but it does corroborate the earlier wall-clock time that wasn't really showing much of a difference.

If we want to add benches to CI later, we probably want to use [lai](https://github.com/bheisler/iai) instead of criterion since that will measure instructions and memory accesses rather than time.