@jdisanti 2021-06-21
Did some more benchmarking since simplifying the benchmarks in #507. It looks like performance is in the same ball park after this PR.

main:
```
deserialization_bench   time:   [12.200 us 12.701 us 13.262 us]
deserialization_bench   time:   [10.637 us 11.046 us 11.456 us]
deserialization_bench   time:   [8.4684 us 8.5261 us 8.5913 us]
deserialization_bench   time:   [8.1389 us 8.2168 us 8.3122 us]
```

PR:
```
deserialization_bench   time:   [11.877 us 11.980 us 12.097 us]
deserialization_bench   time:   [11.694 us 11.765 us 11.846 us]
deserialization_bench   time:   [11.302 us 11.402 us 11.516 us]
deserialization_bench   time:   [11.235 us 11.299 us 11.369 us]
```

It looks like there may be some opportunity for improvement, but we're on the order of a 0-4 microsecond difference. The consistency of measurements of the PR version vs. what's in main is interesting.